All works done.

+++ Study/add_plan.cpp
@@ -4,14 +4,23 @@
 #include <QWidget>
 #include <QString>
 #include <QDebug>
-#include<QComboBox>
-#include<QCalendarWidget>
+#include <QComboBox>
+#include <QCalendarWidget>
+#include <QImage>
+#include <QPalette>
 Add_Plan::Add_Plan(QWidget *parent) :
     QWidget(parent),
     ui(new Ui::Add_Plan)
 {
 
     ui->setupUi(this);
+    QImage _image;
+     _image.load("://imgs/logo.png");
+     setAutoFillBackground(true);   // 这个属性一定要设置
+      QPalette pal(palette());
+       pal.setBrush(QPalette::Window, QBrush(_image.scaled(size(), Qt::IgnoreAspectRatio,
+                         Qt::SmoothTransformation)));
+      setPalette(pal);
     db=QSqlDatabase::addDatabase("QSQLITE");
     db.setDatabaseName("learn.db");
     db.open();
@@ -35,39 +44,44 @@ Add_Plan::Add_Plan(QWidget *parent) :
         ui->stackwidget->setCurrentIndex(0);
         addComboBox();
     });
+    //添加
     connect(ui->btn_add,&QPushButton::clicked,[=](){
         Dialog dialog;
-        dialog.exec();
-        LearnData data;
-        data=dialog.getinput();
-        //qDebug() << data.time;
-        bool flag=1;
-        for(int i=0;i<ui->tableView->model()->rowCount();i++){
-            if(ui->tableView->model()->index(i,0).data().toString()==data.title){
-                flag=0;
-                break;
+        if(dialog.exec()==QDialog::Accepted){
+            LearnData data;
+            data=dialog.getinput();
+            //qDebug() << data.time;
+            bool flag=1;
+            for(int i=0;i<ui->tableView->model()->rowCount();i++){
+                if(ui->tableView->model()->index(i,0).data().toString()==data.title){
+                    flag=0;
+                    break;
+                }
+            }
+            if(flag){
+                QString sql="insert into planinfo(planname,plandetails,time) values('"+data.title+"','"+data.details+"','"+data.time.toString("yyyy-MM-dd")+"')";
+                query->exec(sql);
+                qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
+                setRemind();
+            }
+            else{
+                QMessageBox::critical(this,"添加失败","已存在同名内容");
             }
         }
-        if(flag){
-            QString sql="insert into planinfo(planname,plandetails,time) values('"+data.title+"','"+data.details+"','"+data.time.toString("yyyy-MM-dd")+"')";
-            query->exec(sql);
-            qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
-            setRemind();
-        }
-        else{
-            QMessageBox::critical(this,"添加失败","已存在同名内容");
-        }
+
 
     });
+    //删除
     connect(ui->btn_del,&QPushButton::clicked,[=](){
-//        QString planname = ui->tableView->currentIndex().data().toString();
-        QString sql="delete from planinfo where planname='"+curPlanName+"'" ;
-        QString sql2="delete from planremind where planname='"+curPlanName+"'" ;
-        query->exec(sql);
-        query->exec(sql2);
-        qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
-        ui->textBrowser->setText("<h1 align=center>请选择任务</h1>");
-        setRemind();
+        if(QMessageBox::question(this,"确认删除吗？","确认删除吗？")==QMessageBox::Yes){
+            QString sql="delete from planinfo where planname='"+curPlanName+"'" ;
+            QString sql2="delete from planremind where planname='"+curPlanName+"'" ;
+            query->exec(sql);
+            query->exec(sql2);
+            qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
+            ui->textBrowser->setText("<h1 align=center>请选择任务</h1>");
+            setRemind();
+        }
 
     });
     connect(ui->tableView,&QTableView::clicked,this,[=](const QModelIndex &index){
@@ -80,17 +94,20 @@ Add_Plan::Add_Plan(QWidget *parent) :
                                                  "<p align=right>创建于：%3</p>"
                                                  ).arg(curPlanName).arg(curPlanDetails).arg(curTime));
     });
+    //修改
     connect(ui->btn_change,&QPushButton::clicked,[=](){
         Dialog dialog;
-        dialog.exec();
-        LearnData data;
-        data=dialog.getinput();
-        QString sql="update planinfo set planname='"+data.title+"',plandetails='"+data.details+"',time='"+data.time.toString("yyyy-MM-dd")+"' where planname='"+curPlanName+"'";
-        QString sql2="delete from planremind where planname='"+curPlanName+"'" ;
-        query->exec(sql);
-        query->exec(sql2);
-        qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
-        setRemind();
+        dialog.setContents(curPlanName,curPlanDetails,QDate::fromString(curTime));
+        if(dialog.exec()==QDialog::Accepted){
+            LearnData data;
+            data=dialog.getinput();
+            QString sql="update planinfo set planname='"+data.title+"',plandetails='"+data.details+"',time='"+data.time.toString("yyyy-MM-dd")+"' where planname='"+curPlanName+"'";
+            QString sql2="delete from planremind where planname='"+curPlanName+"'" ;
+            query->exec(sql);
+            query->exec(sql2);
+            qmodel->setQuery("select planname as 任务名称,time as 创建时间,plandetails as 任务内容 from planinfo");
+            setRemind();
+        }
     });
     //子窗口2
     //添加任务
@@ -98,6 +115,20 @@ Add_Plan::Add_Plan(QWidget *parent) :
     //显示任务
     void(QComboBox::*fp)(int)=&QComboBox::currentIndexChanged;
     connect(ui->comboBox,fp,this,[=](){
+        query->exec("select * from planinfo where planname='"+ui->comboBox->currentText()+"'");
+        query->next();
+        QString curPlanName = query->value("planName").toString();
+        QString curPlanDetails = query->value("planDetails").toString();
+        QString curTime = query->value("time").toDate().toString("yyyy-MM-dd");
+
+//        qDebug() <<query->value("planname").toString();
+        ui->textBrowser_2->setText(    QObject::tr("<h1 align=center>%1</h1>"
+                                                 "<p> <font size='4'>&nbsp;&nbsp;%2</font></p>"
+                                                 "<p align=right>创建于：%3</p>"
+                                                 ).arg(curPlanName).arg(curPlanDetails).arg(curTime));
+    });
+    void(QComboBox::*fp2)(int)=&QComboBox::activated;
+    connect(ui->comboBox,fp2,this,[=](){
         query->exec("select * from planinfo where planname='"+ui->comboBox->currentText()+"'");
         query->next();
         QString curPlanName = query->value("planName").toString();
@@ -136,14 +167,6 @@ Add_Plan::Add_Plan(QWidget *parent) :
                                                  "<p align=right>创建于：%3</p>"
                                                  ).arg(curPlanName).arg(curPlanDetails).arg(curTime));
     });
-
-
-
-
-
-
-
-
 
 
     //设置返回按钮
